self-supervised/SupCon/requirements.txt
albumentations==0.5.2
faiss-gpu==1.6.5
matplotlib==3.3.3
MulticoreTSNE==0.1
numpy==1.19.4
opencv-python==4.4.0.46
opencv-python-headless==4.4.0.46
pandas==1.1.5
Pillow==8.0.1
protobuf==3.14.0
pytorch-metric-learning==0.9.95
pytorch-ranger==0.1.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
tensorboard==2.4.0
timm==0.3.4
torch==1.7.1
torch-ema @ git+https://github.com/fadel/pytorch_ema@27afe25b9fb9f0d05a87ae94e4e4ad9e92d70a85
torch-lr-finder==0.2.1
torch-optimizer==0.0.1a17
torchvision==0.8.2
tqdm==4.54.1
cmake==3.18.4




